equivalent tiles: imported tile_import functions
